requirements.txt
torchvision
torch
numpy
tqdm
matplotlib
pyyaml
python-box